Abstract RemoteNodeState HTTP implementation to HttpRemoteNodeState
