+++ presto-main/src/main/java/com/facebook/presto/metadata/DiscoveryNodeManager.java
@@ -177,7 +177,7 @@ public final class DiscoveryNodeManager
         // Add new nodes
         for (InternalNode node : aliveNodes) {
             nodeStates.putIfAbsent(node.getNodeIdentifier(),
-                    new RemoteNodeState(httpClient, uriBuilderFrom(node.getInternalUri()).appendPath("/v1/info/state").build()));
+                    new HttpRemoteNodeState(httpClient, uriBuilderFrom(node.getInternalUri()).appendPath("/v1/info/state").build()));
         }
 
         // Schedule refresh
